+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Transmission Release Notes
 # --------------------------------------------------------------------
+# Date: 21-May-2019
+# Version: 1.0.2
+# Supported API versions: 
+Transmission  v 2.92 (14714)
+# Supported Mule Runtime Versions: 
+3.9.2
+# New Features and Functionality
+Added method `remove-torrent`
+# Closed Issues in this release
+Fixed reconnection after 409 exception
+Fixed field `hashString`
+# Known Issues in this release
+Get torrents doesn't support get torrent by id
+# --------------------------------------------------------------------
 # Date: 30-Jan-2017
 # Version: 1.0
 # Supported API versions: